feat: impplemented authentication user API and hashing password
