+++ src/main/java/com/scaffold/spring_boot/controller/AuthenticationController.java
@@ -1,8 +1,9 @@
 package com.scaffold.spring_boot.controller;
 
+import com.scaffold.spring_boot.dto.request.ApiResponse;
 import com.scaffold.spring_boot.dto.request.AuthenticationRequest;
 import com.scaffold.spring_boot.dto.response.AuthenticationResponse;
-import com.scaffold.spring_boot.service.UserService;
+import com.scaffold.spring_boot.service.AuthenticationService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -14,10 +15,12 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/auth")
 @RequiredArgsConstructor
 public class AuthenticationController {
-    private final UserService userService;
+    private final AuthenticationService authenticationService;
 
-    @PostMapping
-    public AuthenticationResponse authenticate(@RequestBody @Valid AuthenticationRequest authenticationRequest) {
-        return userService.authenticateUser(authenticationRequest);
+    @PostMapping("login")
+    public ApiResponse<AuthenticationResponse> authenticate(@RequestBody @Valid AuthenticationRequest authenticationRequest) {
+        return ApiResponse.<AuthenticationResponse>builder().
+                result(authenticationService.authenticateUser(authenticationRequest))
+                .build();
     }
 }
